No need to hardcode asset paths
