+++ Gruntfile.js
@@ -91,7 +91,7 @@ module.exports = function(grunt) {
             },
             'bootjs': {
                 'files': {
-                    'build/boot.js': ['src/js/boot.js'],
+                    'build/boot.js': ['src/js/boot.js.tpl'],
                 }
             },
             'embed': {
@@ -218,6 +218,7 @@ module.exports = function(grunt) {
                         middlewares.unshift(function (req, res, next) {
                             res.setHeader('Access-Control-Allow-Origin', '*');
                             res.setHeader('Access-Control-Allow-Methods', '*');
+
                             return next();
                         });
                         return middlewares;
@@ -236,6 +237,12 @@ module.exports = function(grunt) {
         });
     });
 
+    grunt.registerTask('loadLocalConfig', function() {
+        grunt.config('visuals', {
+            assetPath: 'http://localhost:8000'
+        });
+    });
+
     grunt.registerTask('boot_url', function() {
         grunt.log.write('\nBOOT URL: '['green'].bold);
         grunt.log.writeln(grunt.template.process('<%= visuals.s3.domain %><%= visuals.s3.path %>/boot.js'));
@@ -247,7 +254,7 @@ module.exports = function(grunt) {
     grunt.registerTask('embed', ['shell:embed', 'template:embed', 'sass:embed']);
     grunt.registerTask('interactive', ['shell:interactive', 'template:bootjs', 'sass:interactive']);
     grunt.registerTask('all', ['interactive', 'embed', 'postcss', 'copy:assets']);
-    grunt.registerTask('default', ['clean', 'copy:harness', 'all', 'connect', 'watch']);
+    grunt.registerTask('default', ['loadLocalConfig', 'clean', 'copy:harness', 'all', 'connect', 'watch']);
     grunt.registerTask('build', ['clean', 'all']);
     grunt.registerTask('deploy', ['loadDeployConfig', 'prompt:visuals', 'build', 'copy:deploy', 'aws_s3', 'boot_url']);
