+++ src/js/boot.js.tpl
@@ -1,0 +1,29 @@
+'use strict';
+define([], function() {
+    function addCSS(url) {
+        var head = document.querySelector('head');
+        var link = document.createElement('link');
+        link.setAttribute('rel', 'stylesheet');
+        link.setAttribute('type', 'text/css');
+        link.setAttribute('href', url);
+        head.appendChild(link);
+    }
+
+    return {
+        boot: function(el, context, config, mediator) {
+            config = {
+                'assetPath': '<%= assetPath %>'
+            };
+
+            // Load CSS asynchronously
+            window.setTimeout(function() {
+                addCSS('<%= assetPath %>/style.css');
+            }, 10);
+
+            // Load JS and init
+            require(['<%= assetPath %>/main.js'], function(main) {
+                main.init(el, context, config, mediator);
+            }, function(err) { console.error('Error loading boot.', err); });
+        }
+    };
+});
